Tweak other window activity colors
Thanks to the currently active window and other windows with activity
colors being very similar, it was often quite hard to see which window
is currently active.

This changes activity color in other windows to just be gray, rather
than the specific color of the theme, allowing the currently active
window to stand out more.